# Desktop/ProyectoDSW/backend/src/main/java/com/venta/backend/cliente/application/dto/response/ClienteResponse.java
package com.venta.backend.cliente.application.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDate;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ClienteResponse {

    private Long clienteId;
    private String dni;
    private String firstName;
    private String lastName;
    private String fullName;
    private String email;
    private String phoneNumber;
    private String address;
    private LocalDate registrationDate;
    private String estado;
    private String categoria;
}